requirements.txt
pillow
pytesseract
pytest
coverage